*created constants fee module
